[JS] Rename bot to agent throughout the project (#119)
* rename bot -> agent

* fix: update ACR login token secret reference in CI workflow

* rename weather-agent

* add node weather-agent to cd

* move workspace under samples

* rename tasks in cd
