--- samples/basic/echo-bot/nodejs/.vscode/launch.json
@@ -1,17 +1,0 @@
-{
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "type": "node",
-            "request": "launch",
-            "name": "node-echo-bot",
-            "skipFiles": [
-                "<node_internals>/**"
-            ],
-            "program": "${workspaceFolder}/dist/index.js",
-            "outFiles": [ "${workspaceFolder}/**/dist/**/*.js" ],
-            "envFile": "${workspaceFolder}/.env",
-            "preLaunchTask": "npm: build echo-bot"
-        }
-    ]
-}
