--- samples/basic/echo-bot/nodejs/.vscode/tasks.json
@@ -1,15 +1,0 @@
-{
-	"version": "2.0.0",
-	"tasks": [
-		{
-			"type": "npm",
-			"script": "build",
-			"group": {
-				"kind": "build",
-				"isDefault": true
-			},
-			"problemMatcher": [],
-			"label": "npm: build echo-bot"
-		}
-	]
-}
