+++ samples/basic/empty-agent/dotnet/AspNetExtensions.cs
