+++ samples/basic/empty-agent/dotnet/ConversationStateExtensions.cs
