+++ samples/basic/empty-agent/dotnet/EchoBot.csproj
