+++ samples/basic/empty-agent/dotnet/MyBot.cs
