+++ samples/basic/empty-agent/dotnet/Properties/launchSettings.TEMPLATE.json
