+++ samples/basic/empty-agent/dotnet/appsettings.json
