+++ samples/basic/empty-agent/nodejs/.npmrc
