+++ samples/basic/empty-agent/nodejs/Dockerfile
