+++ samples/basic/empty-agent/nodejs/env.TEMPLATE
@@ -2,3 +2,4 @@
 tenantId=
 clientId=
 clientSecret=
+DEBUG=agents:*:error
