+++ samples/basic/empty-agent/nodejs/src/agent.ts
