+++ samples/basic/empty-agent/nodejs/src/index.ts
