+++ samples/basic/empty-agent/nodejs/tsconfig.json
