+++ samples/basic/weather-agent/dotnet/Agents/WeatherForecastAgent.cs
