+++ samples/basic/weather-agent/dotnet/Agents/WeatherForecastAgentResponse.cs
