+++ samples/basic/weather-agent/dotnet/AspNetExtensions.cs
