+++ samples/basic/weather-agent/dotnet/MyAgent.cs
