+++ samples/basic/weather-agent/dotnet/Plugins/AdaptiveCardPlugin.cs
