+++ samples/basic/weather-agent/dotnet/Plugins/DateTimePlugin.cs
