+++ samples/basic/weather-agent/dotnet/Plugins/WeatherForecast.cs
