+++ samples/basic/weather-agent/dotnet/Plugins/WeatherForecastPlugin.cs
