+++ samples/basic/weather-agent/dotnet/WeatherBot.csproj
