+++ samples/basic/weather-agent/dotnet/appsettings.json
