+++ samples/basic/weather-agent/nodejs/Dockerfile
