+++ samples/basic/weather-agent/nodejs/env.TEMPLATE
