+++ samples/basic/weather-agent/nodejs/package.json
