+++ samples/basic/weather-agent/nodejs/src/myAgent.ts
