+++ samples/basic/weather-agent/nodejs/src/tools/dateTimeTool.ts
