+++ samples/basic/weather-agent/nodejs/src/tools/getWeatherTool.ts
